frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["Alexander Lyotov <sippie84@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.2"
httpx = "^0.23.0"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
